Feat: Add a pipeline node to the agent canvas
