+++ web/src/constants/agent.ts
@@ -116,4 +116,5 @@ export enum Operator {
   Splitter = 'Splitter',
   HierarchicalMerger = 'HierarchicalMerger',
   Extractor = 'Extractor',
+  Generate = 'Generate',
 }
